refactor: Move markdown rendering functions to dedicated file
- Create pkg/too/output/to_markdown.go with all markdown-specific rendering
- Remove markdown functions from engine.go (reduced by ~85 lines)
- Better separation of concerns - markdown logic is now isolated
